Move things to decoders.rs
